+++ Runtime/Events/DialogueSignal.cs
@@ -1,0 +1,12 @@
+using System;
+using UnityEngine.Events;
+
+namespace Runtime
+{
+    [Serializable]
+    public class DialogueSignal
+    {
+        public string Name;
+        public UnityEvent Event;
+    }
+}